feat: 리크루트 관리 기능을 위한 커스텀 훅 추가 및 관련 페이지 수정
